--- src/main/java/APIFactory.java
@@ -1,6 +1,0 @@
-package io.spicelabs.rodeocomponents;
-
-public interface APIFactory<T extends API> {
-    String name();
-    T buildAPI(RodeoComponent subscriber);
-}
