--- src/main/java/APIFactoryReceiver.java
@@ -1,5 +1,0 @@
-package io.spicelabs.rodeocomponents;
-
-public interface APIFactoryReceiver {
-  public <T extends API> void  publishFactory(RodeoComponent publisher, String apiName, APIFactory<T> apiFactory, Class<T> apiType);
-}
