--- src/main/java/APIS/RodeoArgumentConstants.java
@@ -1,5 +1,0 @@
-package io.spicelabs.rodeocomponents.APIS;
-
-public class RodeoArgumentConstants {
-  public final static String NAME = "ArgumentRegistrar";
-}
