--- src/main/java/APIS/RodeoArgumentListener.java
@@ -1,9 +1,0 @@
-package io.spicelabs.rodeocomponents.APIS;
-
-import java.util.List;
-
-public interface RodeoArgumentListener {
-  RodeoArgumentMemento begin();
-  RodeoArgumentResult receiveArgument(List<String> parameters, RodeoArgumentMemento memento);
-  RodeoArgumentResult end(RodeoArgumentMemento memento);
-}
