--- src/main/java/APIS/RodeoArgumentMemento.java
@@ -1,3 +1,0 @@
-package io.spicelabs.rodeocomponents.APIS;
-
-public interface RodeoArgumentMemento { }
