--- src/main/java/APIS/RodeoArgumentRegistrar.java
@@ -1,7 +1,0 @@
-package io.spicelabs.rodeocomponents.APIS;
-
-import io.spicelabs.rodeocomponents.API;
-
-public interface RodeoArgumentRegistrar extends API {
-    void register(RodeoArgumentListener listener);
-}
