--- src/main/java/APIS/RodeoLoggerConstants.java
@@ -1,5 +1,0 @@
-package io.spicelabs.rodeocomponents.APIS;
-
-public class RodeoLoggerConstants {
-  public static final String NAME = "RodeoLogger";
-}
