--- src/main/java/Releasable.java
@@ -1,5 +1,0 @@
-package io.spicelabs.rodeocomponents;
-
-public interface Releasable {
-  void release();
-}
