--- src/main/java/RodeoComponent.java
@@ -1,5 +1,0 @@
-package io.spicelabs.rodeocomponents;
-
-public interface RodeoComponent extends RodeoComponentv0 {
-  
-}
