--- src/main/java/RodeoIdentity.java
@@ -1,6 +1,0 @@
-package io.spicelabs.rodeocomponents;
-
-public interface RodeoIdentity {
-  String name();
-  String publisher();
-}
