+++ src/main/java/io/spicelabs/rodeocomponents/RodeoComponentv0.java
@@ -1,0 +1,28 @@
+package io.spicelabs.rodeocomponents;
+
+/* Copyright 2026 Stephen Hawley, Spice Labs, Inc. & Contributors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License. */
+
+import java.lang.Runtime.Version;
+
+public interface RodeoComponentv0 {
+  void initialize() throws Exception;
+  RodeoIdentity getIdentity();
+  Version getComponentVersion();
+  void exportAPIFactories(APIFactoryReceiver receiver);
+  void importAPIFactories(APIFactorySource factorySource);
+  void onLoadingComplete();
+  void shutDown();
+}
+
